+++ data/invalid/datapackage_no_foreign_key.json
@@ -1,0 +1,49 @@
+{
+    "name": "testing",
+    "resources": [
+        {
+            "name": "data",
+            "path": "data.csv",
+            "schema": {
+                "fields": [
+                    {
+                        "name": "id",
+                        "type": "string",
+                        "constraints": {
+                            "required": true
+                        }
+                    },
+                    {
+                        "name": "name",
+                        "type": "string"
+                    },
+                    {
+                        "name": "description",
+                        "type": "string"
+                    },
+                    {
+                        "name": "amount",
+                        "type": "number"
+                    }
+                ],
+                "primaryKey": "id"
+            }
+        },
+        {
+            "name": "data2",
+            "path": "data2.csv",
+            "schema": {
+                "fields": [
+                    {
+                        "type": "string",
+                        "name": "parent"
+                    },
+                    {
+                        "type": "string",
+                        "name": "comment"
+                    }
+                ]
+            }
+        }
+    ]
+}
