+++ frictionless/resource/resource.py
@@ -8,22 +8,22 @@ import attrs
 from typing_extensions import Self
 
 from .. import errors, fields, helpers, settings
+from ..checklist import Checklist
 from ..detector import Detector
 from ..dialect import Control, Dialect
 from ..exception import FrictionlessException
 from ..metadata import Metadata
 from ..platform import platform
+from ..report import Report
 from ..schema import Schema
 from ..system import system
-from ..validator import Validator
 from .factory import Factory
 from .stats import ResourceStats
 
 if TYPE_CHECKING:
     from .. import types
-    from ..checklist import Checklist
+    from ..error import Error
     from ..package import Package
-    from ..report import Report
     from ..system import Loader
 
 
@@ -619,8 +619,112 @@ class Resource(Metadata, metaclass=Factory):  # type: ignore
             Report: validation report
 
         """
-        validator = Validator()
-        return validator.validate_resource(self, checklist=checklist)
+        # Create state
+        partial = False
+        timer = helpers.Timer()
+        labels: List[str] = []
+        errors: List[Error] = []
+        warnings: List[str] = []
+
+        # Prepare checklist
+        checklist = checklist or Checklist()
+        checks = checklist.connect(self)
+
+        # Validate metadata
+        try:
+            self.to_descriptor(validate=True)
+        except FrictionlessException as exception:
+            return Report.from_validation_task(
+                self, time=timer.time, errors=exception.to_errors()
+            )
+
+        # TODO: remove in version 6
+        # Ignore not-supported hashings
+        if self.hash:
+            algorithm, _ = helpers.parse_resource_hash_v1(self.hash)
+            if algorithm not in ["md5", "sha256"]:
+                warning = "hash is ignored; supported algorithms: md5/sha256"
+                warnings.append(warning)
+
+        # Prepare resource
+        if self.closed:
+            try:
+                self.open()
+            except FrictionlessException as exception:
+                self.close()
+                return Report.from_validation_task(
+                    self, time=timer.time, errors=exception.to_errors()
+                )
+
+        # Validate data
+        with self:
+            # Validate start
+            for index, check in enumerate(checks):
+                for error in check.validate_start():
+                    if error.type == "check-error":
+                        del checks[index]
+                    if checklist.match(error):
+                        errors.append(error)
+
+            # Validate file
+            if not isinstance(self, platform.frictionless_resources.TableResource):
+                if self.hash is not None or self.bytes is not None:
+                    helpers.pass_through(self.byte_stream)
+
+            # Validate table
+            else:
+                row_count = 0
+                labels = self.labels
+                while True:
+                    row_count += 1
+
+                    # Emit row
+                    try:
+                        row = next(self.row_stream)  # type: ignore
+                    except FrictionlessException as exception:
+                        errors.append(exception.error)
+                        continue
+                    except StopIteration:
+                        break
+
+                    # Validate row
+                    for check in checks:
+                        for error in check.validate_row(row):
+                            if checklist.match(error):
+                                errors.append(error)
+
+                    # Callback row
+                    if on_row:
+                        on_row(row)
+
+                    # Limit rows
+                    if limit_rows:
+                        if row_count >= limit_rows:
+                            warning = f"reached row limit: {limit_rows}"
+                            warnings.append(warning)
+                            partial = True
+                            break
+
+                    # Limit errors
+                    if limit_errors:
+                        if len(errors) >= limit_errors:
+                            errors = errors[:limit_errors]
+                            warning = f"reached error limit: {limit_errors}"
+                            warnings.append(warning)
+                            partial = True
+                            break
+
+            # Validate end
+            if not partial:
+                for check in checks:
+                    for error in check.validate_end():
+                        if checklist.match(error):
+                            errors.append(error)
+
+        # Return report
+        return Report.from_validation_task(
+            self, time=timer.time, labels=labels, errors=errors, warnings=warnings
+        )
 
     # Export
